Moved the downloader and embedded the cache in s3 and github.
